Fix Module C Lifecycle & Saga Reliability and DTO Regressions
- Refactored `HousingTransactionSagaHandler` to use `IHousingService` argument-based methods instead of `dict` hacks, enforcing Protocol Purity.
- Updated `ISimulationState` to include `monetary_ledger: Optional[IMonetaryLedger]`, removing unreliable `getattr` access.
- Updated `IFirm` and `IHousehold` protocols to include `reset()` and `reset_tick_state()`, removing `hasattr` checks in `AgentLifecycleManager`.
- Fixed a critical regression where `DebtStatusDTO` consumers were accessing the deprecated `total_outstanding_debt` float field instead of `total_outstanding_pennies`. Updated `Bank`, `InheritanceManager`, `HousingTransactionHandler`, `Firm`, and `simulation.orchestration.utils` to use the correct integer field and convert where necessary.
- Updated `HousingTransactionSagaHandler` to explicitly convert principal to int pennies before calling `add_lien`, fixing a Financial Integrity violation.
- Updated unit and system tests to align with these architectural hardening changes.

+++ modules/housing/service.py
@@ -47,7 +47,7 @@ class HousingService(IHousingService):
     def transfer_asset(self, asset_id: Any, new_owner_id: Any) -> bool:
         return self.transfer_ownership(int(asset_id), int(new_owner_id))
 
-    def add_lien(self, property_id: int, loan_id: Any, lienholder_id: Optional[int] = None, principal: Optional[float] = None) -> Optional[str]:
+    def add_lien(self, property_id: int, loan_id: Any, lienholder_id: Optional[int] = None, principal: Optional[int] = None) -> Optional[str]:
         # Support for IInventoryHandler signature: add_lien(asset_id, lien_details)
         if isinstance(loan_id, dict) and lienholder_id is None:
              details = loan_id
@@ -75,7 +75,7 @@ class HousingService(IHousingService):
         new_lien: LienDTO = {
             "loan_id": loan_id_str,
             "lienholder_id": int(lienholder_id_val) if lienholder_id_val is not None else -1,
-            "principal_remaining": float(principal_val) if principal_val is not None else 0.0,
+            "principal_remaining": int(principal_val) if principal_val is not None else 0,
             "lien_type": "MORTGAGE"
         }
         unit.liens.append(new_lien)
